test: checkout test with react-testing-library
